+++ Service/src/main/java/com/nf511/flower/service/OrderflowerService.java
@@ -9,7 +9,7 @@ import java.util.List;
 public interface OrderflowerService {
 
     int insertOrderflower(@Param("cartIds") List<Integer> cartIds);
-    int updateOrderflowerBycartId(@Param("orderId") int orderId,@Param("cartIds") List<Integer> cartIds);
+    int updateOrderflowerBycartId(@Param("orderId") long orderId,@Param("cartIds") List<Integer> cartIds);
     List<Orderflower> selectAllByOrderId(@Param("orderId") int orderId);
